added unit tests for param templates create and params template permission

+++ segment/api/views/custom_segment/segment_create_options.py
@@ -60,15 +60,17 @@ class SegmentCreateOptionsApiView(APIView):
 
     def delete(self, request, *args, **kwargs):
         """
-        deletes CTL ParamsTemplate object for a given id
+        deletes CTL ParamsTemplate object for a given id if user is owner
         """
         if request.user.has_permission(StaticPermissions.BUILD__CTL_PARAMS_TEMPLATE):
             template_id = request.data.get("id", None)
             if not isinstance(template_id, int):
-                raise TypeError("id value must be an integer")
+                raise TypeError("Template id value must be an integer.")
             params_template = get_object(ParamsTemplate, id=template_id)
-            params_template.delete()
-            return Response(status=HTTP_200_OK)
+            if params_template.owner.id == request.user.id:
+                params_template.delete()
+                return Response(status=HTTP_200_OK)
+            raise PermissionDenied("Cannot delete a template owned by another user.")
         raise PermissionDenied
 
     def _get_templates(self, segment_type):
